requirements.txt
streamlit
easyocr
opencv-python-headless
pandas
pillow
google-generativeai